+++ src/QuickMarcEditor/QuickMarcEditor.js
@@ -35,6 +35,7 @@ const hotKeys = {
 };
 
 const QuickMarcEditor = ({
+  action,
   instance,
   onClose,
   handleSubmit,
@@ -138,7 +139,15 @@ const QuickMarcEditor = ({
             dismissible
             onClose={onClose}
             defaultWidth="100%"
-            paneTitle={instance ? <FormattedMessage id="ui-quick-marc.record.edit.title" values={instance} /> : ''}
+            paneTitle={instance
+              ? (
+                <FormattedMessage
+                  id={`ui-quick-marc.record.${action}.title`}
+                  values={instance}
+                />
+              )
+              : ''
+            }
             paneSub={(
               <QuickMarcRecordInfo
                 status={initialValues?.updateInfo?.recordState}
@@ -184,6 +193,7 @@ const QuickMarcEditor = ({
 };
 
 QuickMarcEditor.propTypes = {
+  action: PropTypes.string.isRequired,
   instance: PropTypes.object,
   onClose: PropTypes.func.isRequired,
   handleSubmit: PropTypes.func.isRequired,